fix(pretty-format): combine DOMElement plugins
